Refactor code for bettercodehub
Creates a userService and moves two functions into contextpackService to reduce duplicate code
