+++ client/src/app/app.component.ts
@@ -1,7 +1,7 @@
 import { HttpClient } from '@angular/common/http';
 import { Component, OnInit } from '@angular/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { User } from './contextpacks/user';
+import { User } from './user';
 import { SocialAuthService, GoogleLoginProvider, SocialUser } from 'angularx-social-login';
 import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators';
