+++ client/src/app/contextpacks/contextpack-card/contextpack-info.component.ts
@@ -46,44 +46,8 @@ export class ContextPackInfoComponent implements OnInit, OnDestroy {
       this.editField(event[0],event[1],event[2]);
     }
     if(event.length === 2){
-      this.updateField(contextPack,event);
+      this.contextPackService.updateField(contextPack,event);
     }
-  }
-
-  updateField(contextPack: ContextPack, event: string[]): void {
-    //to figure out what field is being changed so the correct http param can be sent
-    let obj: any;
-    switch(event[1]){
-      case 'name':obj =  {name: event[0]};
-        break;
-      case 'enabled':obj =   {enabled: event[0]};
-        break;
-      case 'icon':obj =  {icon: event[0]};
-        break;
-    }
-    this.contextPackService.updateContextPack(contextPack, obj).subscribe(existingID => {
-      this.snackBar.open('Updated field ' + event[1] + ' of pack ' + contextPack.name, null, {
-      duration: 2000,
-    });
-    this.updateLocalFields(contextPack, obj);
-    }, err => {
-      this.snackBar.open('Failed to update the ' + event[1] + ' field with value ' + event[0], 'OK', {
-        duration: 5000,
-      });
-    });
-  }
-
-  updateLocalFields(contextpack: ContextPack, obj: any){
-    if(obj.name){
-      contextpack.name =obj.name;
-    }
-    if(obj.enabled){
-      contextpack.name =obj.name;
-    }
-    if(obj.icon){
-      contextpack.icon = obj.icon;
-    }
-    this.ngOnInit();
   }
 
   editField(list: string, newData: string, field: string){
