+++ client/src/app/contextpacks/contextpack-list/contextpack-list.component.ts
@@ -38,6 +38,10 @@ export class ContextPackListComponent implements OnInit, OnDestroy  {
     });
   }
 
+  updateField(contextPack: ContextPack, event: string[]): void {
+    this.contextpackService.updateField(contextPack, event);
+  }
+
   public updateContextpackFilter(): void {
     this.filteredContextpacks = this.contextpackService.filterContextPacks(
       this.serverFilteredContextpacks, { name: this.contextpackName });
@@ -57,39 +61,4 @@ export class ContextPackListComponent implements OnInit, OnDestroy  {
     }
   }
 
-  updateField(contextPack: ContextPack, event: string[]): void {
-    //to figure out what field is being changed so the correct http param can be sent
-    let obj: any;
-    switch(event[1]){
-      case 'name':obj =  {name: event[0]};
-        break;
-      case 'enabled':obj =   {enabled: event[0]};
-        break;
-      case 'icon':obj =  {icon: event[0]};
-        break;
-    }
-    this.contextpackService.updateContextPack(contextPack, obj).subscribe(existingID => {
-      this.snackBar.open('Updated field ' + event[1] + ' of pack ' + contextPack.name, null, {
-      duration: 2000,
-    });
-    this.updateLocalFields(contextPack, obj);
-    }, err => {
-      this.snackBar.open('Failed to update the ' + event[1] + ' field with value ' + event[0], 'OK', {
-        duration: 5000,
-      });
-    });
-  }
-
-  updateLocalFields(contextpack: ContextPack, obj: any){
-    if(obj.name){
-      contextpack.name =obj.name;
-    }
-    if(obj.enabled){
-      contextpack.name =obj.name;
-    }
-    if(obj.icon){
-      contextpack.icon = obj.icon;
-    }
-    this.ngOnInit();
-  }
 }
